fix format

+++ src/Token.sol
@@ -4,13 +4,11 @@ pragma solidity ^0.8.13;
 import "openzeppelin-contracts/contracts/token/ERC20/ERC20.sol";
 
 contract Token is ERC20 {
-    constructor(string memory name, string memory symbol)
-        ERC20(name, symbol)
-    {
+    constructor(string memory name, string memory symbol) ERC20(name, symbol) {
         _mint(msg.sender, 1e24);
     }
 
-    function mint(address to, uint amount) external {
+    function mint(address to, uint256 amount) external {
         _mint(to, amount);
     }
 }
